Some fixes for the sizes of things being passed over blobs.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@2707 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/analysis/trunk/parallelanalysis/DuchampParallel.cc
@@ -442,7 +442,7 @@ namespace askap {
 
 
 
-            }
+	    }
         }
 
         //**************************************************************//
@@ -1120,7 +1120,7 @@ namespace askap {
                     if (this->itsCube.pars().getFlagATrous()) this->itsCube.ReconCube();
                     else if (this->itsCube.pars().getFlagSmooth()) this->itsCube.SmoothCube();
 
-                    int64 size = this->itsCube.getSize();
+                    int32 size = this->itsCube.getSize();
                     float mean = 0., rms;
                     float *array;
                     // make a mask in case there are blank pixels.
@@ -1192,7 +1192,7 @@ namespace askap {
                 }
 
                 // use it to calculate the rms for this section
-                int64 size = this->itsCube.getSize();
+                int32 size = this->itsCube.getSize();
                 double rms = 0.;
                 float *array;
 
@@ -1242,7 +1242,7 @@ namespace askap {
                 // get the means from the workers
                 ASKAPLOG_INFO_STR(logger,  this->workerPrefix() << "Receiving Means and combining");
                 LOFAR::BlobString bs1;
-                int size = 0;
+                int64 size = 0;
                 double av = 0;
 
                 for (int i = 1; i < itsNNode; i++) {
@@ -1252,7 +1252,7 @@ namespace askap {
                     int version = in.getStart("meanW2M");
                     ASKAPASSERT(version == 1);
                     double newav;
-                    int64 newsize;
+                    int32 newsize;
                     int16 rank;
                     in >> rank >> newav >> newsize;
                     in.getEnd();
@@ -1308,7 +1308,7 @@ namespace askap {
                 // get the means from the workers
                 ASKAPLOG_INFO_STR(logger,  this->workerPrefix() << "Receiving RMS values and combining");
                 LOFAR::BlobString bs;
-                long size = 0;
+                int64 size = 0;
                 double rms = 0;
 
                 for (int i = 1; i < itsNNode; i++) {
@@ -1318,7 +1318,7 @@ namespace askap {
                     int version = in.getStart("rmsW2M");
                     ASKAPASSERT(version == 1);
                     double newrms;
-                    int64 newsize;
+                    int32 newsize;
                     int16 rank;
                     in >> rank >> newrms >> newsize;
                     in.getEnd();
